requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.5
defusedxml==0.6.0
Django==3.1.6
django-allauth==0.44.0
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
httpie==2.4.0
idna==2.10
Markdown==3.3.3
oauthlib==3.1.0
psycopg2-binary==2.8.6
pycparser==2.20
Pygments==2.7.4
PyJWT==2.0.1
PySocks==1.7.1
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
